CORE-1876 Rename `server/amqp` to notifications
